+++ bruges/filters/__init__.py
@@ -4,6 +4,7 @@ from .wavelets import sweep
 from .wavelets import rotate_phase
 
 from .kernels import gaussian_kernel
+from .kernels import gaussian
 
 from .filters import snn
 from .filters import kuwahara
@@ -13,7 +14,7 @@ from .anisodiff import anisodiff, anisodiff3
 
 __all__ = ["ricker", "ormsby", "sweep",
            "rotate_phase",
-           "gaussian_kernel",
+           "gaussian_kernel", "gaussian",
            "snn", "kuwahara", "conservative",
            "anisodiff", "anisodiff3",
            ]