+++ bruges/filters/kernels.py
@@ -8,7 +8,10 @@
 import numpy as np
 
 
-def gaussian_kernel(size, size_y=None):
+def gaussian(size, size_y=None):
+    """
+    Returns a Gaussian kernel.
+    """
     size = int(size)
     if not size_y:
         size_y = size
@@ -17,3 +20,6 @@ def gaussian_kernel(size, size_y=None):
     x, y = np.mgrid[-size:size+1, -size_y:size_y+1]
     g = np.exp(-(x**2/float(size)+y**2/float(size_y)))
     return g / g.sum()
+
+
+gaussian_kernel = gaussian