+++ bruges/reflection/reflection.py
@@ -6,11 +6,14 @@ Various reflectivity algorithms.
 :copyright: 2015 Agile Geoscience
 :license: Apache 2.0
 """
+from collections import namedtuple
+
 import numpy as np
 from numpy import tan, sin, cos
 
 from bruges.rockphysics import moduli
 from bruges.rockphysics import anisotropy
+from bruges.util import deprecated
 
 
 def scattering_matrix(vp1, vs1, rho1, vp0, vs0, rho0, theta1=0):
@@ -203,7 +206,9 @@ def akirichards(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     term4 = -1 * z * sin(theta1)**2
 
     if terms:
-        return term1, term2, term3, term4
+        fields = ['term1', 'term2', 'term3', 'term4']
+        AkiRichards = namedtuple('AkiRichards', fields)
+        return AkiRichards(term1, term2, term3, term4)
     else:
         return (term1 + term2 + term3 + term4)
 
@@ -256,7 +261,9 @@ def akirichards_alt(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     term3 = 0.5 * dvp/vp * (tan(theta)**2 - sin(theta)**2)
 
     if terms:
-        return term1, term2, term3
+        fields = ['term1', 'term2', 'term3']
+        AkiRichards = namedtuple('AkiRichards', fields)
+        return AkiRichards(term1, term2, term3)
     else:
         return (term1 + term2 + term3)
 
@@ -305,7 +312,9 @@ def fatti(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     term3 = -1 * (0.5 * tan(theta1)**2 - 2 * (vs/vp)**2 * sin(theta1)**2) * d
 
     if terms:
-        return term1, term2, term3
+        fields = ['term1', 'term2', 'term3']
+        Fatti = namedtuple('Fatti', fields)
+        return Fatti(term1, term2, term3)
     else:
         return (term1 + term2 + term3)
 
@@ -350,39 +359,31 @@ def shuey(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     term3 = f * (np.tan(theta1)**2 - np.sin(theta1)**2)
 
     if terms:
-        return term1, term2, term3
+        fields = ['R0', 'Rg', 'Rf']
+        Shuey = namedtuple('Shuey', fields)
+        return Shuey(term1, term2, term3)
     else:
         return (term1 + term2 + term3)
 
 
-def shuey2(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
+@deprecated('Please use shuey() instead.')
+def shuey2(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0):
     """
     Compute Shuey approximation with 2 terms.
-
-    Wraps shuey().
     """
-    r0, rg, rf = shuey(vp1, vs1, rho1, vp2, vs2, rho2, theta1, terms=True)
-
-    if terms:
-        return r0, rg
-    else:
-        return r0 + rg
+    r, g = shuey(vp1, vs1, rho1, vp2, vs2, rho2, theta1=theta1, terms=True)[:2]
+    return r + g
 
 
+@deprecated('Please use shuey() instead.')
 def shuey3(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     """
-    Compute Shuey approximation with 2 terms.
-
-    Wraps shuey().
+    Compute Shuey approximation with 3 terms.
     """
-    r0, rg, rf = shuey(vp1, vs1, rho1, vp2, vs2, rho2, theta1, terms=True)
-
-    if terms:
-        return r0, rg, rf
-    else:
-        return r0 + rg + rf
+    return shuey(vp1, vs1, rho1, vp2, vs2, rho2, theta1=theta1)
 
 
+@deprecated('Please use bortfeld() instead.')
 def bortfeld2(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     """
     The 2-term Bortfeld approximation for ava analysis.
@@ -418,7 +419,12 @@ def bortfeld2(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
         return (term1 + term2)
 
 
+@deprecated('Please use bortfeld() instead.')
 def bortfeld3(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
+    return bortfeld(vp1, vs1, rho1, vp2, vs2, rho2, theta1=theta1)
+
+
+def bortfeld(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     """
     Compute Bortfeld approximation with three terms.
     http://sepwww.stanford.edu/public/docs/sep111/marie2/paper_html/node2.html
@@ -455,14 +461,18 @@ def bortfeld3(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     term3 = 0.5 * dvp/vp * np.tan(theta1)**2 * np.sin(theta1)**2
 
     if terms:
-        return term1, term2, term3
+        fields = ['term1', 'term2', 'term3']
+        Bortfeld = namedtuple('Bortfeld', fields)
+        return Bortfeld(term1, term2, term3)
     else:
         return (term1 + term2 + term3)
 
 
 def hilterman(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     """
-    Hilterman (1989) approximation.
+    Not recommended, only seems to match Zoeppritz to about 10 deg.
+
+    Hilterman (1989) approximation from Mavko et al. Rock Physics Handbook.
     According to Dvorkin: "arguably the simplest and a very convenient
     [approximation]." At least for small angles and small contrasts.
 
@@ -481,16 +491,22 @@ def hilterman(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
              value corresponding to each angle.
     """
     theta1 = np.radians(theta1)
+
     ip1 = vp1 * rho1
     ip2 = vp2 * rho2
-    rpp0 = (ip2 - ip1) / (ip2 + ip1)
-    dpr = moduli.pr(vp2, vs2) - moduli.pr(vp1, vs1)
+    rp0 = (ip2 - ip1) / (ip2 + ip1)
 
-    term1 = rpp0 * np.cos(theta1)**2
-    term2 = 2.25 * dpr * np.sin(theta1)**2
+    pr2, pr1 = moduli.pr(vp2, vs2), moduli.pr(vp1, vs1)
+    pravg = (pr2 + pr1) / 2.
+    pr = (pr2 - pr1) / (1 - pravg)**2.
+
+    term1 = rp0 * np.cos(theta1)**2.
+    term2 = pr * np.sin(theta1)**2.
 
     if terms:
-        return term1, term2
+        fields = ['term1', 'term2']
+        Hilterman = namedtuple('Hilterman', fields)
+        return Hilterman(term1, term2)
     else:
         return (term1 + term2)